updated build script

+++ .build.sh
@@ -9,14 +9,19 @@ port=8083
 
 echo "restart production server"
 nodeid=`ps -aux | grep --extended-regexp "[${port:0:1}]${port:1}" | awk '{print $2}'`
-kill $nodeid
+
+if [ $nodeid ]
+then
+    kill $nodeid
+fi
+
 cd $production
 git pull origin master
 (node $main $port &)
 
 echo "move compressed files and sha/gpg signatures to packages directory"
-git archive --format=tar -v -o $project.tar.gz HEAD
-git archive --format=zip -v -o $project.zip HEAD
+git archive --format=tar -o $project.tar.gz HEAD
+git archive --format=zip -o $project.zip HEAD
 
 sha256sum *.tar.gz *.zip > sha256sums.txt
 gpg --pinentry-mode loopback --passphrase $gpgpass --batch --yes --detach-sign -a sha256sums.txt
